test: 큐 소비자 테스트 및 EventListener 테스트 수정 (#60)
